requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
neo4j>=5.12.0
qdrant-client>=1.6.4
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-environ>=0.11.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3
django-import-export==4.3.3
drf-yasg